+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.10.3]
 - add range support to match (not variable)
 - add pipemethod support to `|>`
+- fix range add overflow
 
 ## [0.10.2]
 - add `_` as blank expression